feat: 구역, 업무 이름이 상단에 노출되도록 수정
